feat(testnet status): refactor
